requirements.txt
crewai
crewai-tools
python-dotenv
yfinance
litellm
